feat(mps-model-adapters): remove code for mps-model-server-plugin
BREAKING CHANGE: We decided to remove the components light-model-client and mps-model-sever-plugin.
  The light-model-server/client was not used required continuous maintenance.
   As an alternative, we suggest using the mps-sync-plugin to sync models with a model server and access the data through the model-sever.
  See https://issues.modelix.org/issue/MODELIX-996
Fixes: MODELIX-996
