+++ commitlint.config.js
@@ -8,8 +8,6 @@ module.exports = {
         "bulk-model-sync",
         "bulk-model-sync-gradle",
         "deps",
-        "light-model-client",
-        "model-server-lib",
         "metamodel-export",
         "model-api-gen",
         "model-api-gen-gradle",
@@ -20,7 +18,6 @@ module.exports = {
         "modelql",
         "mps-model-adapters",
         "mps-model-server",
-        "mps-model-server-plugin",
         "openapi",
         "ts-model-api",
         "vue-model-api",
