--- light-model-client/src/commonMain/kotlin/org/modelix/client/light/ReadWriteLock.kt
@@ -1,8 +1,0 @@
-package org.modelix.client.light
-
-internal expect class ReadWriteLock() {
-    fun <T> runRead(body: () -> T): T
-    fun <T> runWrite(body: () -> T): T
-    fun canRead(): Boolean
-    fun canWrite(): Boolean
-}
