--- light-model-client/src/jsMain/kotlin/org/modelix/client/light/WebsocketConnectionJS.kt
@@ -1,31 +1,0 @@
-/*
- * Copyright (c) 2022.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.modelix.client.light
-
-import io.ktor.client.HttpClient
-import io.ktor.client.engine.js.Js
-import io.ktor.client.plugins.websocket.WebSockets
-import org.modelix.model.server.api.ModelQuery
-
-fun createLightModelClient(url: String = "ws://localhost:48302/ws", query: ModelQuery? = null): LightModelClient {
-    val httpClient = HttpClient(Js) {
-        install(WebSockets)
-    }
-    val modelClient = LightModelClient.builder().httpClient(httpClient).url(url).build()
-    if (query != null) modelClient.changeQuery(query)
-    return modelClient
-}
