+++ model-api-gen-gradle-test/kotlin-generation/src/test/kotlin/org/modelix/modelql/typed/TypedModelQLTest.kt
@@ -30,12 +30,19 @@ import jetbrains.mps.core.xml.C_XmlDocument
 import jetbrains.mps.core.xml.C_XmlFile
 import jetbrains.mps.lang.editor.imageGen.C_ImageGenerator
 import jetbrains.mps.lang.editor.imageGen.ImageGenerator
+import kotlinx.coroutines.runBlocking
+import org.junit.jupiter.api.TestInstance
+import org.modelix.apigen.test.ApigenTestLanguages
 import org.modelix.metamodel.instanceOf
 import org.modelix.metamodel.typed
 import org.modelix.metamodel.untyped
+import org.modelix.model.ModelFacade
 import org.modelix.model.api.INode
 import org.modelix.model.api.remove
 import org.modelix.model.api.resolve
+import org.modelix.model.client2.ModelClientV2
+import org.modelix.model.client2.runWrite
+import org.modelix.model.lazy.RepositoryId
 import org.modelix.modelql.client.ModelQLClient
 import org.modelix.modelql.core.asMono
 import org.modelix.modelql.core.count
@@ -66,6 +73,7 @@ import org.modelix.modelql.gen.jetbrains.mps.lang.editor.imageGen.setNode
 import org.modelix.modelql.untyped.children
 import org.modelix.modelql.untyped.conceptReference
 import org.modelix.modelql.untyped.descendants
+import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNotEquals
@@ -73,9 +81,35 @@ import kotlin.test.assertNotNull
 import kotlin.test.assertNull
 import kotlin.test.assertTrue
 
-abstract class TypedModelQLTest {
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+class TypedModelQLTest {
 
-    abstract fun runTest(block: suspend (ModelQLClient) -> Unit)
+    private var testRun = 0
+    private val modelClient = ModelClientV2.builder().url("http://localhost:28102/v2/").build().also { runBlocking { it.init() } }
+    private val branchRef
+        get() = ModelFacade.createBranchReference(RepositoryId("modelql-test$testRun"), "master")
+
+    private fun runTest(block: suspend (ModelQLClient) -> Unit) {
+        val modelQlClient = ModelQLClient.builder()
+            .url("http://localhost:28102/v2/repositories/${branchRef.repositoryId.id}/branches/${branchRef.branchName}/query")
+            .build()
+
+        runBlocking { block(modelQlClient) }
+    }
+
+    init {
+        ApigenTestLanguages.registerAll()
+    }
+
+    @BeforeTest
+    fun setup() {
+        testRun++
+        runBlocking {
+            modelClient.runWrite(branchRef) {
+                createTestData(it)
+            }
+        }
+    }
 
     protected fun createTestData(rootNode: INode) {
         rootNode.allChildren.forEach { it.remove() }
