--- mps-model-server-plugin/build.gradle.kts
@@ -1,66 +1,0 @@
-import org.modelix.mpsHomeDir
-
-plugins {
-    `modelix-kotlin-jvm`
-    alias(libs.plugins.intellij)
-    `modelix-project-repositories`
-}
-
-dependencies {
-    implementation(project(":model-server-lib"))
-    implementation(project(":mps-model-adapters"))
-}
-
-// Configure Gradle IntelliJ Plugin
-// Read more: https://plugins.jetbrains.com/docs/intellij/tools-gradle-intellij-plugin.html
-intellij {
-    localPath = mpsHomeDir.get().asFile.absolutePath
-    instrumentCode = false
-}
-
-tasks {
-    patchPluginXml {
-        sinceBuild.set("203")
-        untilBuild.set("241.*")
-    }
-
-    buildSearchableOptions {
-        enabled = false
-    }
-
-//    signPlugin {
-//        certificateChain.set(System.getenv("CERTIFICATE_CHAIN"))
-//        privateKey.set(System.getenv("PRIVATE_KEY"))
-//        password.set(System.getenv("PRIVATE_KEY_PASSWORD"))
-//    }
-//
-//    publishPlugin {
-//        token.set(System.getenv("PUBLISH_TOKEN"))
-//    }
-
-    runIde {
-        autoReloadPlugins.set(true)
-    }
-
-    val mpsPluginDir = project.findProperty("mps.plugins.dir")?.toString()?.let { file(it) }
-    if (mpsPluginDir != null && mpsPluginDir.isDirectory) {
-        create<Sync>("installMpsPlugin") {
-            dependsOn(prepareSandbox)
-            from(project.layout.buildDirectory.dir("idea-sandbox/plugins/mps-model-server-plugin"))
-            into(mpsPluginDir.resolve("mps-model-server-plugin"))
-        }
-    }
-}
-
-group = "org.modelix.mps"
-
-publishing {
-    publications {
-        create<MavenPublication>("maven") {
-            artifactId = "model-server-plugin"
-            artifact(tasks.buildPlugin) {
-                extension = "zip"
-            }
-        }
-    }
-}
