+++ settings.gradle.kts
@@ -14,7 +14,6 @@ include("bulk-model-sync-lib")
 include("bulk-model-sync-lib:mps-test")
 include("bulk-model-sync-mps")
 include("kotlin-utils")
-include("light-model-client")
 include("metamodel-export")
 include("model-api")
 include("model-api-gen")
@@ -26,7 +25,6 @@ include("model-datastructure")
 include("model-server")
 include("model-server-openapi")
 include("model-server-api")
-include("model-server-lib")
 include("modelql-client")
 include("modelql-core")
 include("modelql-html")
@@ -35,6 +33,5 @@ include("modelql-typed")
 include("modelql-untyped")
 include("mps-model-adapters")
 include("mps-model-adapters-plugin")
-include("mps-model-server-plugin")
 include("ts-model-api")
 include("vue-model-api")
